test: 분리된 각 AudioService 객체 UnitTest 수정
- Playable, Recordable 추상화 없이 바로 테스트
